Extract index to isolate module
store: extract index specific stuff to isolate module;
various: various refactors and very minor fixes.
