+++ perun/view/flow/run.py
@@ -29,6 +29,7 @@ def process_title(ctx, _, value):
       Func of 'of-key' through 'through-key' for each 'by-key' (stacked)
 
     :param click.Context ctx: called context of the process
+    :param object _: unused parameter
     :param object value: value that is being processed ad add to parameter
     :returns object: either value (if it is non-None) or default title of the graph
     """
